Added simulation of subsampling of cache traces in resmgrreplay
Added the possibility to subsample a complete trace in resmgrreplay according to the method presented in Waldspurger, C. A., Park, N., Garthwaite, A., & Ahmad, I. (2015). Efficient MRC construction with SHARDS. In FAST 15.
Subsampling the cache trace according to this method should enable to simulate the impact of different cache sizes or checkpoint depth from a much smaller trace without lose much accuracy. This change in resmgrreplay helps to measure how much accuracy can be lost due to subsampling.

[Substrate:8a9ee7ee7971f6ad705449d3e8defef80cd1dfb4]