Adding subsampling of cache traces in ESE
Added the possibility to trace a subset of pages in cache according to the method presented in Waldspurger, C. A., Park, N., Garthwaite, A., & Ahmad, I. (2015). Efficient MRC construction with SHARDS. In FAST 15.
This change makes possible to collect enough data to be able to simulate and study cache performance (for example with difference parameters such as the cache size) with a much lower overhead since only a few pages are traced.
The subsampling is activated only if the keyword 'BFRESMGR' is set when collecting the trace, and the ratio of sampled pages can be set with the parameter 'JET_paramFlight_CacheTraceSamplingRatio'.

[Substrate:217dee6e6b8e0d3732d4b7a1a661f01928cc6353]