ESE Block Cache:  try to reduce slab access overhead in RequestFinalizeRead
In prod we are still spending a huge amount of time in RequestFinalizeRead accessing slabs.  This is true even though we don't cache reads.  We also don't seem to be hitting the cache very much because ErrVerifyCluster is a very small percentage of the cost.

This change causes each CRequest to remember if it had a cache hit or miss.  We can then use the cache hit flag to rule out accesses to the cache.  This could allow us to avoid accessing the slab.  This is more certain than using the presence filter.  If this does work then we may need to look more into the effectiveness of the presence filter.

[Substrate:7464b6e75b808cf65e14ae076908f3474fd8ec5f]