ESE Block Cache:  allow uncached file access to succeed if the underlying file does not support FileIdInfo
During testing, it was discovered that accessing files in certain cases (e.g. \\tsclient\c\foo.txt) failed because GetFileInformationByHandleEx with FileIdInfo was failing with ERROR_INVALID_PARAMETER when the file id was being queried for the file while it was being opened.

This change traps this case and causes an invalid file id to be returned.  This case was already established to be valid for a file that cannot ever participate in EBC caching.

[Substrate:b8ce56bd1da9fef79a88d092da7520abce640908]