ESE Block Cache:  fix Cuckoo Filter issue with standby list
The cuckoo filter reference implementation contained a standby list to hold any item that couldn't fit in the main table.  I implemented this because this is the kind of thing that isn't added without a good reason.

The problem is that while we faithfully add/remove items from the standby list, we forget to consult it on the set membership check!  So we can return a false negative for an item in the standby list which is not allowed.

This fix consults the standby list during the set membership check.

[Substrate:dbc2e5d499d2ff297f07ce3ea28b59cd6c202cb1]