ESE Block Cache:  fix lost write caused by incorrect update of presence filter during eviction
When we load a slab for the first time, we enumerate the entries and load the presence filter appropriately.  On any update, we analyze the update and make a relative change to the presence filter.  Because of how this works, we can only consider the contents of each updated slot to make the decision to add or remove entries from the presence filter.  It turned out that tracking any "clean" block or the latest version of a "dirty" block was practical.  When a block is evicted, we will remove it from the PF if it was a "clean" block or the current "dirty" block.  When a block is cached, we will add it to the PF if it was a "clean" block or the first "dirty" of a block.  This may result in two PF entries for the same block if we cached the "clean" page on a read miss and then dirtied it (until that obsolete image is evicted).  This was a compromise because there is no state to determine that the first "dirty" of the block overrides a "clean" block in our current scheme.

The bug was that my code to find the "clean" block was slightly incorrect.  It recognized any clean block including one that was previously dirty, had been written back, and was now being evicted.  I was testing !FDirty when I should have been testing !FDirty && !FEverDirty.

This change also includes:
- some cleanup to ErrLoad to clarify what it is doing
- more fixes to the presence filter related to the standby list
  - remove wasn't updating the standby list
  - retiring a standby item had bad flow control
- moved the call to Update from ErrUpdateSlabs to ErrScheduleSlabForWriteBack to clarify/guarantee it is always called for any slab update and document that it is only done at do time not redo time

Other changes:
- synctool.bat updated to find Product Studio binary that was moved

[Substrate:2488950c19bed3c44f9cdbc0cf8ec27319035f66]