ESE Block Cache:  avoid corruption if we ever access cached file offsets >= 16TB
When inspecting the code, I noticed that we have a potential corruption issue related to integer truncation.

The HashedLRUKCache caches files using 4kb blocks.  It only reserves 32 bits to store the block number.  It also reserves 2^32-1 as an invalid block number.  If the accessed offset of a file / 2^12 >= 2^32-1 (i.e. about 16TB) then the cache truncates the block number and Bad Things will happen.

This change addresses this by preventing caching of these offsets.  It doesn't seem likely at this time that we will have dbs in prod that are this large so this is fine for now.

[Substrate:f8b8ec23f64da9e64e78651ac496ccf2c7f36366]