ESE:  add optional type check to read shadowed header
When the EBC master enable was switched off, we tried to run recovery against a cached EDB file.  This resulted in a recovery failure but with the weird error JET_errPageSizeMismatch.  We should have failed with JET_errFileInvalidType instead.

This change adds an optional type check to ErrUtilReadShadowedHeader etc.  The caller must indicate which type of file they are attempting to access.  If the file may be of any type then JET_filetypeUnknown may be passed.  We will also allow JET_filetypeTempDatabase if JET_filetypeDatabase was indicated (i.e. accept the more specialized type).

This change also cleaned up the cut&paste code duplication in ErrUtilReadSpecificShadowedHeader.

[Substrate:af3a3ef50f46ad12f310a0aeaad5e471fdc95d2f]