ESE:  simplify hung io detection in BF to avoid accessing invalid IO contexts
BF has code to detect hung IOs. It does this by capturing the pvIOContext provided by IFileAPI::PfnIOHandoff and later uses this to call IFileAPI::DtickIOElapsed. The problem with this code is that it can use the pvIOContext after it has been released by the IFileAPI implementation.  Specifically, this happens because BF captures the pvIOContext in calls surrounding sync I/O calls such that there is a timing window where the pvIOContext has been released but it hasn't been removed from the associated BF.

For COSFile, this is a minor annoyance because pvIOContext is an IOREQ and these are never released. Worst case, BF will check the wrong request. If it does, chances are excellent it won't be hung because obviously it was just reused.

For the ESE Block Cache, this is a major annoyance. EBC's file system filter had to build an expensive scheme to validate these requests and return a safe answer if the CIOComplete had been released already. We had to pay this penalty even for files that were not attached to a cache. Even a probabilistic scheme like using a TPool<CIOComplete> with a min lifetime of 10s was not enough to avoid this hazard.

This change addresses the root cause of this problem by eliminating the concept of a pvIOContext for sync I/O.  A NULL will always be returned by IFileAPI::PfnIOHandoff for a sync I/O.  BF has been changed to compensate for this by complicating the usage of BF::pvIOContext.  For an async I/O it stores the pvIOContext from IFileAPI::PfnIOHandoff.  For a sync I/O, it stores a TICK.  These are distinguished by the least significant bit, which is set for a TICK and clear for a pvIOContext.  BF uses the TICK to directly track the elapsed time of sync I/Os. EBC's expensive compensation scheme for this hazard was also removed and replaced with a pooled allocation for CIOComplete.

[Substrate:bb29f616fd6a88422f6f2e74b690ab5b8d039b2f]