ESE Block Cache:  copy status
This change exposes the caching state of an EDB file to copy status.

This is implemented by tagging the EDB file with a particular alternate data stream indicating the caching status.  This can then be easily checked for existence even if the EDB file is in use by another process.  This state is queried via a helper method in BlockCache.Interop which is being (directly) used by Repl for the first time.

[Substrate:43eb6c8a2e01ac5e6b7534ba5bdf1b5264829985]