ESE Block Cache:  Perf:  flush all state on a clean dismount to prevent repeated recovery on file accesses
Perf data from prod shows that HA file checker is consuming significant CPU just accessing the database file header for a cached file.  This is because the cache is not in a clean shutdown state and so it must recompute significant state on mount.

This change causes us to attempt a best effort flush of all state (to the caching file, i.e. flash) on a clean dismount of the cache.  This should be much cheaper than running recovery.

[Substrate:2172d622faf9571b8e953d74524f6f2e3e1eee7a]