ESE:  JetRenameColumn errata
JetRenameColumn has bad error handling where if an error occurs after it commits its trx then it will still try to rollback.  This used to be silently ignored but now it causes an Enforce failure.  This bad error handling has been fixed.

JetRenameColumn does not work properly on a versioned column (e.g. recently added).  Currently, we will crash with the above error.  Previously, we would succeed in updating the catalog but the TDB wasn't updated and we would either return JET_errColumnNotFound or crash.  We now explicitly reject this case with JET_errIllegalOperation.

[Substrate:d01bbfebd903ad15e2618e0fa2a1663492891bb7]