ESE Block Cache:  fix cache failure under high load (such as eseutil /r)
For a while, I have been seeing stress failures where the cache fails to clean a slab during a normal update because all of the blocks in that slab were subject to active IO Range Locks.  I knew that this was likely due to a large IO Range Lock of 3MB vs the 16MB test cache size but I didn't know exactly what was happing.  It was rare so it was a low priority to fix.

Recently, I was using eseutil /r in prod to reproduce a transient lost flush bug.  I found that eseutil /r would very easily hit this same condition.  The normal passive/active copy activity has never hit this condition even once.  Eseutil /r is old code that was written w/o regard for sharing the hardware so obviously increased aggressiveness was a factor.

To attack this problem, I added a bunch of instrumentation for IO Range Locks in the context of overall usage, usage per thread, and usage per slab.  The usage per slab is the key metric that can cause problems for the clean process.

In this process I discovered that we were not honoring the IO aggregation limits.  This was causing potentially huge IOs to be accumulated and these caused single huge IO Range Locks to be acquired.  This is the cause of the stress failure.

Of course, it isn't enough to just prevent these from being aggregated.  We must also determine when "too much" IO is outstanding and pause until we can do more to avoid excessive locking.  We use the per slab IO Range Lock stats to determine when this happens and we will avoid issuing more IO until we are back under the limits.  If the IO is sync or a thread hasn't yet issued an async IO then we will always allow one IO to proceed to guarantee forward progress.

Note that this is primarily a hazard with the hash function.  In prod, the 15GB cache I was testing has 35k slabs each holding 13 database pages.  We were unlucky enough to have enough outstanding IO Range Locks to cover all 13 pages on one of them.  This is why this change uses the per slab stats to achieve this goal rather than some static limit per thread or cache.

Other fixes:
-  Clean up of TFileFilter's IO combination logic where we were taking the grbitQOS from the wrong source
-  Reduced some Persistence Filter tracing

[Substrate:3cdf9fd9140ded493d2d9cdf2eab82bd4a17e5ce]