ESE Block Cache:  Write Caching Percentage fixes
For test purposes, EBC is run locally with "accept forever onetest stress" using the test configuration where all ESE files are cachable.  We also configure the cache to be 50/50 read/write to exercise read caching.  This differs from the current deployment in prod where we are at 0/100 read/write caching.

I never really fully tested this part.  While debugging issues where the stress test failed with JET_errInternalError when cleaning slabs during read/write operations, I discovered that this write caching percentage logic was broken.  Basically, we would still fail to clean the slab in cases that actually should have worked because of this broken logic trying to protect the caching percentage.  For example, we would refuse to cache a read because we thought we didn't have enough write cache but we really had plenty.

We care about fixing this because it makes the stress test run longer without failure and it also is preparation for a likely future where both read and write caching are enabled in production.

This change makes the following improvements which I think solve almost all of the ErrCleanSlab failure cases:

- the logic in CCleanSlabVisitor related to counting what is in the read or write portion of the cache (i.e. clean blocks vs. updated blocks, as indicated by FEverDirty) was fixed to not count superceded pages and to count pages produced rather than visited

-  CCleanSlabVisitor was improved to
   -  properly account for pages that are in the read or write portion of the cache (FValid, !FSuperceded, reads are !FEverDirty and writes are FEverDirty)
   -  properly measure read/write cache blocks produced rather than just visited
   -  identify the read/write cache blocks that are not protected by the write caching percentage for either read or write operations so that they can be written back and/or evicted
   -  failing that, identify cases where a request (read or write) is ok because it won't change the write caching percentage.  for example, a read can evict another read and a write can evict another write
   -  failing that, if the caller says to ignore the write caching percentage then evict anyways
   -  clean attempts that fail due to write caching percentage are now traced
   -  we no longer try to get the IO Range Lock for superceded blocks because it isn't necessary
   -  clean attempts that fail due to IO Range Lock failures have better tracing
-  ErrCleanSlab was improved to ignore (but emit an event) for a failure to cache reads as this isn't fatal
-  ErrCleanSlab was improved to retry on a failure to cache a write using the new flag to ignore write caching percentage because a failure must cause the write request to fail.  this would only help for cases where there is read cache that was protected by the write caching percentage and not protected by an IO Range Lock

[Substrate:66e142039555482b9e2626cfb3cdc1710b2df4ce]