ESE Block Cache:  Stability:  FlushAllState on dismount crashes due to previous state
The FlushAllState that we now due on a prepare to dismount is accidentally touching deallocated state (TLS).

This code path has been modified to clearly separate the HandleJournalFull case and the PrepareToDismount case and the latter case now doesn't touch any of the sync state used by HandleJournalFull.  This should avoid the hazard.

[Substrate:e4758bc5c839ac06d94398fbfafb1c5f97c5fce5]