ESE Block Cache:  Perf:  compress journal entries with LZ4
In prod, we use a significant amount of CPU compressing journal entries.  We are currently using the implementation of Xpress Huffman built into the NT RTL.

This change adds a mode to compress journal entries with the existing LZ4 library included in ESE.  This library is faster and has better compression.

With Xpress Huffman, our compressed journal entries had a bi-modal size of about 2200b and 3500b, as some include one slab of one chunk and others are two slabs of one chunk.  This is a compression ratio of about 0.54 and 0.42.

With LZ4, our compressed journal entries all end up to be about 2300b regardless of the slab count.  This is a compression ratio of 0.56 and 0.28.  These cases are about equal so the overall compression is better.  The average size for LZ4 is about 81% of the size of Xpress Huffman.

We will actually enable the use of LZ4 in a later checkin after this one saturates.  We are doing this to avoid killing database copies if this build rolls back because the downlevel code won't know how to decode journal entries compressed with LZ4.

[Substrate:b403edaec94ed07cdcf207a4411b54fb39fa7c87]