ESE Block Cache:  Perf:  Extend IFileAPI::ErrFlushFileBuffers with FileFlushMode to enable data only flushes
In prod, we are spending a lot of CPU and IO repeatedly flushing the NTFS journal whenever we call ErrFlushFileBuffers on the Caching File or the Cached File.  FlushFileBuffers will cause NTFS to write out the latest state of the file's metadata even if the update is trivial (e.g. Last Access Time).  This is pointless for the EBC scenario.  It is also the only cost here because the actual SYNC command is short circuited because the write cache is disabled on the storage.

The lowest risk fix for this is to use the new (to Win10 RS1) flag to NtFlushBuffersFileEx called FLUSH_FLAGS_FILE_DATA_SYNC_ONLY.  This flag causes any file data to be flushed but skips meta-data changes and also causes a SYNC command to be sent when needed.  This flag is implemented by the OS in such a way that if it isn't recognized then we fall back to the legacy behavior.

IFileAPI::ErrFlushFileBuffers now takes a FileFlushMode with two settings:  ffmAll (the current behavior) and ffmDataOnly (the new behavior).

EBC has been modified to use the new ffmDataOnly mode.

This API is documented here:  https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/ntifs/nf-ntifs-ntflushbuffersfileex

[Substrate:c0d4e863da254a2502d1187b42749b80c4e96811]