Added control of ESE's cache trace subsampling
Added the keyword BFRESMGRSUBSAMPLING to control the subsampling of ESE's cache traces.

Traces are subsampled only if the keyword BFRESMGRSUBSAMPLED is set but keywords BFRESMGR, BF, DataWorkingSet, and Performance are not. Moreover, if the sampling ratio is set to 0, no trace is emitted in order to trace only specific databases.

[Substrate:cfdd1f5d7bd2377a54bf2a104e52cb20b937a728]