ESE:  Stability:  fix hang during sync scavenge
If we are doing a sync clean and things get so bad that we are at max scavenge priority then we will start to issue async write back IO at qosIODispatchImmediate.  The problem with this is that these requests can block rather than be rejected with errDiskTilt.  Normally, if we get errDiskTilt, we will call CBFIssueList::ErrIssue to release any pending work to avoid a hang.  In this case, we don't do that so a hang can occur.

The fix is to detect when we are escaling to qosIODispatchImmediate and call CBFIssueList::ErrIssue first before requesting the async write back.

This was found while running retail stress with EBC enabled.

[Substrate:017b2a98ec34e430b305868f30d1ec729caed277]