ESE Block Cache:  Async IO Worker hang
This change contains a rework of the state management of the cache's CRequest's IO.  It improves the tracking of explicit state related to IO requests and completions.  It also improves the contract with which these are used to perform IO in the multiple rounds we can do (request IO and finalization IO).  This should improve debuggability and increase determinism.  I don't think this was the root bug however.

This change also contains a rework of the sync mechanism used to invoke async IO workers.  After the previous change, I did catch a case where we clearly failed to signal the async IO worker properly.  This could have been the cause of the hang in prod because a missed signal could leave that async IO work stranded indefinitely.  The new scheme uses an atomic update of a single word to control requests to the async IO worker and when that worker needs to run.  This scheme should be more reliable than the current scheme.  It should also fix some performance issues that caused us to use too many worker pool threads.

[Substrate:3f0564af4cee7dc8c7e2ad4213c79877f8a01069]