ESE Block Cache:  fix the durability of IFileAPI::ErrFlushFileBuffers
The current working theory is that this lost write is actually a lost reverted page that happened due to a sequence like this:

- Begin revert and proceed up to where we revert pages
- Revert pages in the EDB
- Flush the FM
- Flush the EDB
- Update the persisted revert checkpoint to JET_revertstateCopingLogs
- HA process restarts
  - some of the page reverts are rolled back when the EBC mounts
- Restart and complete the revert from JET_revertstateCopingLogs
- Recovery begins
- Recovery fails with a lost write

The flush of the EDB maps to a call to IFileAPI::ErrFlushFileBuffers which EBC maps to a call to THashedLRUKCache<ICache>::ErrFlush.  This is currently implemented as a call to flush the EBC's journal.

An EBC journal entry must be explicitly indicated as durable by another subsequent journal entry to be guaranteed to be durable.  This is achieved by flushing the journal again because the subsequent journal entry will indicate that the previous one was considered durable.  If this doesn't happen then the journal entry can be lost on the next mount.  We believe that this is what happened in this case.

The fix is to change THashedLRUKCache<ICache>::ErrFlush to flush the EBC journal twice.

This change also changes ErrPrepareToDismount to use ErrFlush directly which effectively checks that it is working properly.  If it doesn't then we will end up emitting telemetry (HashedLRUKCacheFlushAllStateIncomplete) when the file is closed.  We also fix ErrPrepareToDismount's journal truncation to also be durable for write back to guarantee that we don't redo recovery of state that should already be flushed.

[Substrate:4f6edd6e4b0d9b4a606c398335845fbd2d40b2b1]