ESE Block Cache:  Async IO Worker hang (redo)
Redo of !2247766

In prod, we are seeing occasional hangs in the async io worker state machine. There isn't enough state to debug exactly what happened.

This change contains a rework of the state management of the cache's CRequest's IO. It improves the tracking of explicit state related to IO requests and completions. It also improves the contract with which these are used to perform IO in the multiple rounds we can do (request IO and finalization IO). This should improve debuggability and increase determinism.  This also fixes a race with the completion mechanism based on CMeteredSection by replacing it with a simpler mechanism based on a count.  The race could allow us to touch a released CRequest or even a released cache TLS.

This change also contains a rework of the sync mechanism used to invoke async IO workers. After the previous change, I did catch a case where we clearly failed to signal the async io worker properly. This could have been the cause of the hang in prod because a missed signal could leave that async IO work stranded indefinitely. The new scheme uses an atomic update of a single word to control requests to the async io worker and when that worker needs to run. This scheme should be more reliable than the current scheme. It should also fix some performance issues that caused us to use too many worker pool threads.

[Substrate:dade0d4b554f4928c889f7da2a61fe0c6e01d65f]