Delete RCE in logical rather than allocation order
On passives, RCEs can be allocated out of order because of deferred RCEs which are only allocated later if needed. Cleaning them up in that allocation order can result in RCE chains during the cleanup which are illegal. So, need to clean them up in logical order.
Also added some asserts to make sure deferred RCEs can only be used while in the initial required range.

[Substrate:c84b5929fb7e2e358ed7d8a2fbc84b9dcb4ee810]