ESE: fix ancient problem where we don't EnterDML during schema upgraders
Recently we hit an assert where the FCB hash table was expanding and failed an assert that an entry in the table matched its key.  This happened during a store.worker schema upgrader.  This revealed that we have an ancient issue where if someone is modifying a template table via JET_bitPermitDDL then we should be protecting accesses to the table schema with EnterDML.  This is currently skipped for tables with fixed schema like template tables.

The fix is to EnterDML for template tables if their schema is not truly fixed.  We detect this via the existing FTemplateStatic().  Note that this results in EnterDML being required for updates to template tables which is supported but very rarely used.

This change also significantly improves the UnpublishedJetConvertTests.

[Substrate:a7ac027d4d3439608c08509b7ddfab8faa817d92]