ESE Block Cache: Perf: optimize use of Journal Slabs
In prod, we are spending a lot of time manipulating journal slabs during the evict process.  We are using them to allocate clusters to replace clusters that hold evicted data to support our recovery scheme.  This process is effectively a FIFO allocation scheme and so it should be cheap.  We actually end up using a lot of CPU doing this because we scan the entire slab twice to find the next cluster.  In prod, we end up calling ICachedBlockSlab::ErrGetSlotForWrite 16x per database page.

This change optimizes this case as follows:
- each slab now tracks the count of invalid slots and the lowest invalid slot number
- each journal slab now uses the cached lowest invalid slot to directly find the next cluster
- ErrEvictOrInvalidateSlot uses the count of invalid slots to determine if it needs to flush to reuse the clusters in a journal slab
- ErrEvictOrInvalidateSlot no longer calls ICachedBlockSlab::ErrGetSlotForWrite twice for every operation

Other changes:
- We now use a CJournalSlabWrapper to manage the ref count on the journal slabs, cleaning up some code.
- The alloc of CCachedBlockSlab and CCachedBlockSlabReference and CJournalSlabWrapper are pooled
-  CHashedLRUKCacheThreadLocalStorage::FOpenSlab and FAnyOpenSlab were changed to use the ibSlab as the key to allow wrappers of ICachedBlockSlab like CJournalSlabWrapper to be used

[Substrate:678929bcce24d55713a1ebb161e23f41c2aabc8a]