ESE: Removing EBC tracing calls.
Deactivated the tracing of EBC events that were using the same ETW events as the BFResMgr.
This is required in order to use the sub sampling of the cache traces: EBC produces significantly more events than a subsampled BfResMgr, which produces large trace files, defeating the point of subsampling.

Subsampling could be added to EBC tracing later.

[Substrate:55439d4abe0f32d6b53c2b1dd0930e6f7b0b559c]