Add casts and some #defines preliminary to running GCR2
When GCR2 is run and we collapse all "long"s to "int"s, we end up needing some casts to make the OS happy.  Making a small number of those casts here.  They're NO-OPs in the current code.

Also, first example of the #define foolery we need to make DWORD work.  DWORD in Windows is based on long.  If I was being harsh, I'd remove all usages of DWORD from our code.  But, DWORD is too deep in our psyches to do that.  So, GCR2 will redefine DWORD to be based on "int" and then we have to play a #define game to be able to include Windows headers without type redefinition errors.

Far in the future I see some reorganization of the #include order so that windows specific headers are included in far fewer places.

[Substrate:4463f3c49fff90ba1ded53f1ba56886ea5cb6ad8]