ESE Block Cache: Perf: Combine slab updates for evict and cache
Currently, EBC is written such that if a sync evict is needed to cache something then that evict is processed and journaled separately.  The subsequent update to cache something is its own journal entry.  This is obviously not optimal because we end up processing and journaling the same slab twice per update.

This change reorganizes the code slightly so that these updates are combined.  Now, instead of seeing one journal entry evicting an entry and another one caching an entry in its place, this will be combined into one journal entry where it looks like we just replace one entry with another.

To aid in analysis and debugging, we will also emit the before image of any evicted slot in the combined journal entry.  This slot will be applied to the target slab but will be immediately overwritten with the subsequently added after image.

This impacts several key code paths that analyzed the changes in a slab on update:
-  CUpdateSlabVisitor verifies that the cluster is replaced on any evict/invalidate as this is required to support our recovery scheme
-  CUpdateSlabVisitor collects slots that are involved in the update
-  CCachedBlockPresenceFilter::FUpdateSlot updates the presence filter based on evict/cache as required to correctly identify items that may reside in the cache
-  LogCacheUpdateJournalEntry => TraceCacheUpdateJournalEntry emits a trace on every cache update at do or redo time
-  ErrDumpJournalEntry dumps each entry in the journal

These code paths were updated to use more nuanced checks to still allow the detection of an evict/cache even if it happened to the same slot in one atomic operation.

Some other fixes were made:
-  CCachedBlock::Updno() was exposed to be public because it is used by the new logic above to detect an evict/cache in the same slot using the tuple VolumeId, FileId, FileSerial, CachedBlockNumber, and UpdateNumber to uniquely identify a version of a cached file block
-  We defined an updnoFirst and explicitly prevent it from being used because there is logic in CCachedBlockPresenceFilter that relies on this value uniquely identifying the first update of an item

[Substrate:fc1a83b964e76297b144b8b011919a1bed0e285c]